Break down global typeDefs
